R5-G1-3 (clang-tidy wave 2): qualified auto, redundant casts, std::min/max
Second remediation wave, restricted to purely additive / non-logic checks after
a batch run including readability-simplify-boolean-expr was found to silently
delete a guard (see below and the SAFE-FIX PROTOCOL in
docs/CODEX_REVIEW_5_REMEDIATION.md). This wave clears readability-qualified-auto,
readability-redundant-casting and readability-use-std-min-max across 25 files:

  - qualified-auto adds the pointer '*' to 'auto' iterating raw pointers.
  - redundant-casting drops static_cast<qsizetype>(container.size()), which is
    already qsizetype -- a genuine no-op.
  - use-std-min-max rewrites 'if (a > b) a = b;' as 'a = std::min(a, b);' and the
    max form; the clamp result is identical.

Two edits from this run were reverted by hand as unsafe:
  - redundant-casting removed static_cast<qsizetype> from the fs-tree node-count
    fail-closed guard in partition_apfs_writer.cpp, where the operand is a
    uint64_t -- dropping the cast flips a signed/unsigned comparison on a
    corruption guard. The cast is restored.
  - The whole preceding batch (which also ran simplify-boolean-expr and
    use-ranges) was discarded: simplify-boolean-expr had replaced the body of
    collectExistingFullFsTree with 'return !!collectFullFsTree(...)', deleting
    the create-or-replace name-collision preflight. It compiled; only an
    orphaned-parameter warning-as-error exposed it.

The full diff was read for any removed conditional, loop, or guard before
building. Full Release build + ctest 225/225.

+++ src/core/connectivity_tester.cpp
@@ -189,8 +189,7 @@ void populateMtrResult(MtrResult& result,
     for (int i = 0; i < limit; ++i) {
         result.hops.append(hopStats[i]);
     }
-    result.totalCycles =
-        cancelled ? 0 : static_cast<int>(result.hops.isEmpty() ? 0 : result.hops.first().sent);
+    result.totalCycles = cancelled ? 0 : (result.hops.isEmpty() ? 0 : result.hops.first().sent);
 }
 }  // namespace
